LLMProject/smart-librarian/requirements.txt
chromadb==0.5.5
openai>=1.30.1
python-dotenv>=1.0.1
tqdm>=4.66.0
streamlit>=1.36.0
streamlit-mic-recorder==0.0.8
